[+] configs update

+++ .kodxsetup.sh
@@ -41,17 +41,14 @@ usage() {
     cat <<EOM
     kodx dotfiles setup
     Usage:
-
     install dotfiles and applications
     $ INSALL=1 $0
     or
     $ $0 insall
-
     install dotfiles
     $ INSDOTFILES=1 $0
     or
     $ $0 insdotfiles
-
     install apps
     $ INSAPPS=1 $0
     or
@@ -66,6 +63,11 @@ if [ "$#" -eq 0 ] && [ ! $INSDOTFILES ] && [ ! $INSAPPS ] && [ ! $INSALL ] ; the
 fi
 
 insdotfiles() {
+    echo 'installing dotfiles...'
+    if $(! command -v git > /dev/null); then
+        echo 'git not found, abort'
+        return
+    fi
     local DFDIR="$HOME/.dotfiles"
     [ -d $DFDIR ] && rm -rf $DFDIR
     cd $HOME
@@ -85,7 +87,7 @@ getapp() {
     if [ "$1" = 'curl' ]; then
         GETOPT='-#L -o'
     else
-        GETOPT='--hsts-file /dev/null -nv -O'
+        GETOPT='--hsts-file /dev/null -nv -N -O'
     fi
     local BINPATH="$HOME/bin"
     local CMDPATH="$BINPATH/$(basename $2)"
@@ -95,10 +97,21 @@ getapp() {
     chmod u+rx $CMDPATH
 }
 
+getcmd() {
+    if $(command -v 'wget' > /dev/null); then 
+        echo 'wget'
+    elif $(command -v 'curl' > /dev/null); then 
+        echo 'curl'
+    else 
+        echo ''
+    fi
+}
+
 insapps() {
-    local GETCMD=$(if $(command -v 'curl' > /dev/null); then echo 'curl'; elif $(command -v 'wget' > /dev/null); then echo 'wget'; else echo ''; fi)
+    echo 'installing applications ...'
+    local GETCMD=getcmd
     if [ -z $GETCMD ]; then
-        echo 'curl or wget not found, exit'
+        echo 'wget or curl not found, abort'
         return
     fi
     # set urls to apps
@@ -118,4 +131,5 @@ else
     [ $INSAPPS ] || [ "$1" = 'insapps' ] && insapps
 fi
 
-unset insdotfiles insapps usage getapp
+unset insdotfiles insapps usage getapp getcmd
+
